Cleanup compiler warnings when OPT_WARNINGS is enabled - 'core' and 'include' folders.

+++ core/action_stack.c
@@ -442,7 +442,7 @@ static void action_stack_process()
         // Get function address and id from stack
         long id = action_top(0);
         action_func f = (action_func)action_top(1);
-        if (id == AS_FUNC_ENTRY)    // Safety check
+        if (id == (long)AS_FUNC_ENTRY)    // Safety check
         {
             process = f();
         }
